+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/vite/preprocessor/rewrite.ts
@@ -44,6 +44,9 @@ export const rewrite = ({
 
 				//Add the new spread attribute
 				code.appendRight(element.start + element.name.length + 1, " " + newSpreadAttributeString)
+
+				//add <!-- svelte-ignore a11y-missing-attribute --> just before the element
+				code.appendLeft(element.start, "<!-- svelte-ignore a11y-missing-attribute -->")
 			} else {
 				for (const element_translations of Object.entries(translations)) {
 					const attribute_translations = element_translations[1]
@@ -116,6 +119,9 @@ export const rewrite = ({
 			const newSpreadAttributeString = c.spreadAttr(value)
 
 			code.appendRight(element.start + element.name.length + 1, " " + newSpreadAttributeString)
+
+			//add <!-- svelte-ignore a11y-missing-attribute --> just before the element
+			code.appendLeft(element.start, "<!-- svelte-ignore a11y-missing-attribute -->")
 		} else {
 			for (const [element_name, attribute_translations] of Object.entries(translations)) {
 				for (const { attribute_name, lang_attribute_name } of attribute_translations) {