+++ inlang/source-code/paraglide/paraglide-js/src/compiler/compileMessage.ts
@@ -107,7 +107,7 @@ const messageIndexFunction = (args: {
 	return `/**
  * This message has been compiled by [inlang paraglide](https://inlang.com/m/gerre34r/library-inlang-paraglideJs).
  *
- * - Don't edit the message's code. Use the [inlang ide extension](https://inlang.com/m/r7kp499g/app-inlang-ideExtension),
+ * - Don't edit the message's code. Use [Sherlock (VS Code extension)](https://inlang.com/m/r7kp499g/app-inlang-ideExtension),
  *   the [web editor](https://inlang.com/m/tdozzpar/app-inlang-finkLocalizationEditor) instead, or edit the translation files manually.
  * 
  * - The params are NonNullable<unknown> because the inlang SDK does not provide information on the type of a param (yet).