+++ inlang/source-code/message-lint-rules/emptyPattern/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/message-lint-rule-empty-pattern
 
+## 1.4.4
+
+### Patch Changes
+
+- @inlang/message-lint-rule@1.4.4
+
 ## 1.4.3
 
 ### Patch Changes