+++ inlang/source-code/paraglide/paraglide-js-adapter-unplugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @inlang/paraglide-js-adapter-unplugin
 
+## 1.4.6
+
+### Patch Changes
+
+- Updated dependencies [f6ec6cdc9]
+  - @inlang/paraglide-js@1.2.7
+  - @inlang/sdk@0.26.5
+
+## 1.4.5
+
+### Patch Changes
+
+- Updated dependencies [960f8fb70]
+  - @inlang/paraglide-js@1.2.6
+  - @inlang/sdk@0.26.4
+
 ## 1.4.4
 
 ### Patch Changes