+++ inlang/source-code/plugins/m-function-matcher/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inlang/plugin-m-function-matcher
 
+## 0.8.2
+
+### Patch Changes
+
+- @inlang/sdk@0.26.5
+- @inlang/plugin@2.4.7
+
+## 0.8.1
+
+### Patch Changes
+
+- 960f8fb70: rename the vscode extension to "Sherlock"
+- Updated dependencies [960f8fb70]
+  - @inlang/plugin@2.4.6
+  - @inlang/sdk@0.26.4
+
+## 0.8.0
+
+### Minor Changes
+
+- 92f7c9532: improve matcher to include function referencing & improve test coverage
+
 ## 0.7.3
 
 ### Patch Changes