+++ inlang/source-code/versioned-interfaces/plugin/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @inlang/plugin
 
+## 2.4.7
+
+### Patch Changes
+
+- Updated dependencies [0c272619a]
+  - @inlang/project-settings@2.3.0
+
+## 2.4.6
+
+### Patch Changes
+
+- 960f8fb70: rename the vscode extension to "Sherlock"
+
 ## 2.4.5
 
 ### Patch Changes